refactor(job): modularize job system into registry, loader, and system files
